client side implementation

+++ Shared/Square.cs
@@ -5,8 +5,8 @@ namespace Shared
     public class Square
     {
         public Guid Id { get; set; }
-        public double Top { get; set; }
+        public int Top { get; set; }
 
-        public double Left { get; set; }
+        public int Left { get; set; }
     }
 }
